requirements.txt
mysqlclient>=1.3
requests>=2.0
python-dotenv>=0.6
pandas>=0.20.3
numpy>=1.13.1
scipy>=0.19.1
matplotlib>=2.0.2
